prepare for deploy

+++ db/migrate/20140205182124_drop_user_roles.rb
@@ -1,5 +1,5 @@
 class DropUserRoles < ActiveRecord::Migration
   def change
-    drop_table :user_roles 
+    drop_table :user_roles if table_exists?(:user_roles)
   end
 end
